feat: new user details component
